+++ presentation/src/main/java/com/stop/ui/routedetail/RouteDetailFragment.kt
@@ -6,15 +6,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.view.GravityCompat
-import androidx.fragment.app.activityViewModels
 import androidx.navigation.findNavController
-import com.stop.R
 import androidx.navigation.fragment.findNavController
-import androidx.navigation.navGraphViewModels
 import com.stop.R
+import androidx.navigation.navGraphViewModels
 import com.stop.databinding.FragmentRouteDetailBinding
 import com.stop.domain.model.route.tmap.custom.Coordinate
-import com.stop.ui.route.RouteViewModel
 import com.stop.ui.route.RouteResultViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -23,10 +20,9 @@ class RouteDetailFragment : Fragment(), RouteDetailHandler {
     private var _binding: FragmentRouteDetailBinding? = null
     private val binding get() = _binding!!
 
-    private val routeViewModel: RouteViewModel by activityViewModels()
+    private val routeResultViewModel: RouteResultViewModel by navGraphViewModels(R.id.route_nav_graph)
 
     private lateinit var tMap: RouteDetailTMap
-    private val routeResultViewModel: RouteResultViewModel by navGraphViewModels(R.id.route_nav_graph)
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -47,14 +43,15 @@ class RouteDetailFragment : Fragment(), RouteDetailHandler {
     }
 
     override fun alertTMapReady() {
-        tMap.drawRoutes(routeViewModel.tempItinerary.routes)
+        routeResultViewModel.itinerary.value?.let {
+            tMap.drawRoutes(it.routes)
+        }
     }
 
     private fun initBinding() {
         binding.lifecycleOwner = viewLifecycleOwner
         binding.routeResultViewModel = routeResultViewModel
-        binding.routeViewModel = routeViewModel
-        binding.itinerary = routeViewModel.tempItinerary
+        binding.itinerary = routeResultViewModel.itinerary.value
     }
 
     private fun initTMap() {
@@ -76,7 +73,10 @@ class RouteDetailFragment : Fragment(), RouteDetailHandler {
         }
 
         binding.imageViewClose.setOnClickListener {
-            binding.root.findNavController().navigate(R.id.action_routeDetailFragment_to_mapFragment)
+            findNavController().apply {
+                setGraph(R.navigation.nav_graph)
+                navigate(R.id.action_global_mapFragment)
+            }
         }
     }
 
@@ -89,7 +89,7 @@ class RouteDetailFragment : Fragment(), RouteDetailHandler {
         })
 
         binding.routeDetailDrawer.recyclerViewRouteDetail.adapter = adapter
-        adapter.submitList(routeViewModel.getRouteItems())
+        adapter.submitList(routeResultViewModel.getRouteItems())
     }
 
     override fun onDestroyView() {
